move StoreProvider inside Store
