+++ common/v2/services/index.ts
@@ -4,5 +4,4 @@ export * from './ShapeShift';
 export * from './useStateReducer';
 export * from './Store';
 export { RatesContext, RatesProvider } from './RatesProvider';
-export { StoreContext, StoreProvider } from './StoreProvider';
 export { DevModeContext, DevModeProvider, useDevMode } from './DevModeProvider';
